Create index.html
A simple yet stylish webpage that can sell your music as a producer. Remain anonymous and allow users to sample a few seconds of music before purchasing. 
The page view is responsive and be able to change from list to grid format when displaying the music. 
The page uses mixtures of purple and black and contains a QR code to a link page where social media platforms can be added at a later time.